web: Library store refactoring to support filtering
