+++ packages/web/src/queries/library.ts
@@ -10,10 +10,14 @@ export type getLibraryResponse = {
   };
 };
 
-export async function fetchAll() {
-  const res = await axios.get<getLibraryResponse>(`${library()}?sort=artist`, {
-    headers: { 'Content-Type': 'application/json' },
-  });
+export async function fetch(filter?: string) {
+  const filterString = filter ? `&subkeyfilter=${filter}` : '';
+  const res = await axios.get<getLibraryResponse>(
+    `${library()}?sort=artist${filterString}`,
+    {
+      headers: { 'Content-Type': 'application/json' },
+    },
+  );
 
   const tracks = res.data.data.songs;
 
@@ -25,6 +29,6 @@ export function fetchOne(trackId: string) {
 }
 
 export const libraryApi = {
-  fetchAll,
+  fetch,
   fetchOne,
 };
