+++ packages/web/src/store/slices/library/adapter.ts
@@ -1,0 +1,5 @@
+import { createEntityAdapter } from '@reduxjs/toolkit';
+
+import { Track } from '~/types/library';
+
+export const libraryAdapter = createEntityAdapter<Track>();
