+++ packages/web/src/store/slices/library/library.ts
@@ -1,28 +1,9 @@
-import {
-  createAsyncThunk,
-  createDraftSafeSelector,
-  createEntityAdapter,
-  createSelector,
-  createSlice,
-  EntityId,
-  SerializedError,
-} from '@reduxjs/toolkit';
-
-import { libraryApi } from '~/queries/library';
-import { RootState } from '~/store/store';
-import { Track } from '~/types/library';
-import { stringSanitizer } from '~/utils/stringSanitizer';
-import { getArrayString } from '~/utils/trackMeta';
+import { createSlice, EntityId, SerializedError } from '@reduxjs/toolkit';
 
 import { playerActions } from '../player';
-
-export const fetchLibrary = createAsyncThunk('/library/fetchAll', async () => {
-  const tracks = await libraryApi.fetchAll();
-
-  return tracks;
-});
-
-const libraryAdapter = createEntityAdapter<Track>();
+import { libraryAdapter } from './adapter';
+import { fetchLibrary } from './thunks';
+import { filterTracksByValue } from './utils';
 
 export type LibraryState = {
   loading: boolean;
@@ -44,25 +25,6 @@ const initialState = libraryAdapter.getInitialState<LibraryState>({
   history: [],
 });
 
-function filterTracksByValue(
-  filter: LibraryState['filter'],
-  tracks: Track[],
-  trackIDs: EntityId[],
-) {
-  if (filter === '') return trackIDs;
-
-  return tracks
-    .filter((t) =>
-      (['name', 'artist', 'album'] as (keyof Track)[]).some((p) =>
-        stringSanitizer(
-          // @ts-expect-error
-          Array.isArray(t[p]) ? getArrayString(t?.[p]) : t[p],
-        ).includes(stringSanitizer(filter)),
-      ),
-    )
-    .map((t) => t.id);
-}
-
 export const librarySlice = createSlice({
   name: 'library',
   initialState,
@@ -77,7 +39,11 @@ export const librarySlice = createSlice({
     // library reducers
     builder.addCase(fetchLibrary.fulfilled, (state, { payload }) => {
       state.loading = false;
-      libraryAdapter.upsertMany(state, payload);
+      if (payload.filter) {
+        libraryAdapter.setAll(state, payload.tracks);
+      } else {
+        libraryAdapter.upsertMany(state, payload.tracks);
+      }
     });
     builder.addCase(fetchLibrary.pending, (state) => {
       state.loading = true;
@@ -120,67 +86,5 @@ export const librarySlice = createSlice({
     });
   },
 });
-
-const {
-  selectById: selectTrackById,
-  selectIds: selectTrackIds,
-  selectEntities: selectTrackEntities,
-  selectAll: selectAllTracks,
-  selectTotal: selectTotalTracks,
-} = libraryAdapter.getSelectors<RootState>((state) => state.library);
-
-const selectLibraryLoading = (state: RootState) => state.library.loading;
-const selectLibraryFilter = (state: RootState) => state.library.filter;
-const selectLibraryQueue = (state: RootState) => state.library.queue;
-const selectLibraryActiveTrackIndex = (state: RootState) =>
-  state.library.activeTrackIndex;
-
-const selectFilteredTrackIds = createDraftSafeSelector(
-  selectLibraryFilter,
-  selectAllTracks,
-  selectTrackIds,
-  filterTracksByValue,
-);
-const selectFilteredTrackIdCount = createSelector(
-  selectFilteredTrackIds,
-  (filteredTrackIds) => filteredTrackIds?.length ?? 0,
-);
-const selectLibraryActiveTrackId = createSelector(
-  selectLibraryActiveTrackIndex,
-  selectLibraryQueue,
-  (index, queue) => (index !== null ? queue[index] : null),
-);
-const selectLibraryActiveTrack = createSelector(
-  selectLibraryActiveTrackId,
-  (state) => (id: EntityId) => selectTrackById(state, id),
-  (trackId, tracksById) => {
-    if (trackId === null) return null;
-
-    const track = tracksById(trackId);
-
-    if (!track) return null;
-
-    return track;
-  },
-);
-
-export const librarySelectors = {
-  // entity selectors
-  selectTrackById,
-  selectTrackIds,
-  selectTrackEntities,
-  selectAllTracks,
-  selectTotalTracks,
-  // normal selectors
-  selectLibraryLoading,
-  selectLibraryFilter,
-  selectLibraryQueue,
-  selectLibraryActiveTrackIndex,
-  // memoized selectors
-  selectFilteredTrackIds,
-  selectLibraryActiveTrackId,
-  selectLibraryActiveTrack,
-  selectFilteredTrackIdCount,
-};
 
 export const libraryActions = librarySlice.actions;
